Four Seasons Wonderland
Our Unity3D project is called "Seasons of Wonderland" and it is a collaboration between a team of three people. This project is an adventure game with four scenes representing different seasons, and the main menu scene is the starting point of the game.Players can choose to start the game in the main menu scene, and then enter any one of the four scenes to start their adventure.